requirements.txt
asgiref==3.4.1
astroid==2.8.4
autopep8==1.6.0
azure-core==1.19.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-storages==1.12.3
idna==3.3
isodate==0.6.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.1.1
packaging==21.0
Pillow==8.4.0
pip-review==1.1.0
pkg_resources==0.0.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.2
